requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.7
django-bootstrap-v5==1.0.0
djangorestframework==3.12.2
gunicorn==20.0.4
importlib-metadata==2.1.1
numpy==1.20.1
Pillow==8.1.2
psycopg2==2.8.6
pymongo==3.11.3
pytz==2021.1
redis==3.5.3
soupsieve==2.2
sqlparse==0.4.1
zipp==3.4.1
